Adapt the tests to pass.
The clang based parser can not work with incomplete sources. When it sees the adoption of a protocol, it must have seen the declaration of that protocol before. For real life code, that happens through #include and #import statements in the preamble of each file, but for these tests the code fragments needed some additions.